NIFI-4730 restructure of nifi-gcp-bundle
This closes #2367.
