+++ nifi-nar-bundles/nifi-gcp-bundle/nifi-gcp-processors/pom.xml
@@ -38,7 +38,12 @@
             <groupId>org.apache.nifi</groupId>
             <artifactId>nifi-processor-utils</artifactId>
         </dependency>
-        <dependency>
+         <dependency>
+            <groupId>org.apache.nifi</groupId>
+            <artifactId>nifi-gcp-services-api</artifactId>
+            <scope>provided</scope>
+        </dependency>
+       <dependency>
             <groupId>org.apache.nifi</groupId>
             <artifactId>nifi-mock</artifactId>
             <scope>test</scope>
@@ -78,23 +83,6 @@
             <artifactId>json</artifactId>
             <version>1.8</version>
         </dependency>
-        <dependency>
-            <groupId>com.google.auth</groupId>
-            <artifactId>google-auth-library-oauth2-http</artifactId>
-            <version>0.6.0</version>
-            <exclusions>
-                <exclusion>
-                    <groupId>com.google.code.findbugs</groupId>
-                    <artifactId>jsr305</artifactId>
-                </exclusion>
-            </exclusions>
-        </dependency>
-        <dependency>
-            <groupId>com.github.stephenc.findbugs</groupId>
-            <artifactId>findbugs-annotations</artifactId>
-            <version>1.3.9-1</version>
-        </dependency>
-
     </dependencies>
 
     <build>
